chore: refactor, split files, remove dependency gonfig
