+++ go.mod
@@ -6,12 +6,10 @@ require (
 	github.com/Tedyst/sqlitestore v0.0.0-20190801085407-9d526ef60b1c
 	github.com/koesie10/webauthn v0.3.5-0.20190427130103-61460705ee9a
 	github.com/mattn/go-sqlite3 v1.10.1-0.20190716071503-d6f416f91c46
-	github.com/tkanos/gonfig v0.0.0-20181112185242-896f3d81fadf
 	go.uber.org/zap v1.10.0
 )
 
 require (
-	github.com/ghodss/yaml v1.0.0 // indirect
 	github.com/gorilla/securecookie v1.1.1 // indirect
 	github.com/gorilla/sessions v1.2.0 // indirect
 	github.com/pkg/errors v0.8.0 // indirect
@@ -21,5 +19,4 @@ require (
 	go.uber.org/multierr v1.1.0 // indirect
 	golang.org/x/crypto v0.0.0-20181024171144-74cb1d3d52f4 // indirect
 	gopkg.in/square/go-jose.v2 v2.1.9 // indirect
-	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
